+++ custom-elements-manifest.config.mjs
@@ -13,6 +13,7 @@ fs.mkdirSync('dist', { recursive: true });
 export default {
   globs: ['src/components/**/cc-*.js'],
   exclude: ['src/**/*.stories.js'],
+  litelement: true,
   // dev: true,
   // watch: true,
   plugins: [